qwen_v2509/requirements.txt
diffusers==0.35.2
transformers==4.57.3
accelerate==1.12.0
xformers
Pillow
runpod==1.7.13
requests
